requirements.txt
Flask==2.2.3
spacy==3.5.1
transformers==4.26.1
torch>=1.10.0
scikit-learn==1.2.1
PyPDF2==3.0.1
python-dotenv==0.21.0
pytest==7.2.2
pytest-cov==4.0.0
requests==2.28.2
gunicorn==20.1.0